% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_anograms_wide.R
\docType{data}
\name{anograms_wide}
\alias{anograms_wide}
\title{Anograms Wide Example data}
\format{A data frame with 3 rows and 12 variables:
\describe{
  \item{none_5}{none_5}
  \item{none_6}{none_6}
  \item{none_7}{none_7}
  \item{none_8}{none_8}
  \item{alone_5}{alone_5}
  \item{alone_6}{alone_6}
  \item{alone_7}{alone_7}
  \item{alone_8}{alone_8}
  \item{withEgo_5}{withEgo_5}
  \item{withEgo_6}{withEgo_6}
  \item{withEgo_7}{withEgo_7}
  \item{withEgo_8}{withEgo_8}
}}
\usage{
anograms_wide
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
