% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_audience_wide.R
\docType{data}
\name{audience_wide}
\alias{audience_wide}
\title{Audience Wide Example data}
\format{A data frame with 12 rows and 4 variables:
\describe{
  \item{id}{id}
  \item{one}{one}
  \item{twenty}{twenty}
  \item{large}{large}
}}
\usage{
audience_wide
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
