% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_brain_wide.R
\docType{data}
\name{brain_wide}
\alias{brain_wide}
\title{Brain Wide Example data}
\format{A data frame with 6 rows and 9 variables:
\describe{
  \item{left_digit}{left_digit}
  \item{left_letter}{left_letter}
  \item{left_mixed}{left_mixed}
  \item{right_digit}{right_digit}
  \item{right_letter}{right_letter}
  \item{right_mixed}{right_mixed}
  \item{none_digit}{none_digit}
  \item{none_letter}{none_letter}
  \item{none_mixed}{none_mixed}
}}
\usage{
brain_wide
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
