% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_child_variables.R
\docType{data}
\name{child_vars}
\alias{child_vars}
\title{Child Variables Example data}
\format{A data frame with 8 rows and 4 variables:
\describe{
  \item{child}{child}
  \item{shoe}{shoe}
  \item{read}{read}
  \item{age}{age}
}}
\usage{
child_vars
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
