% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coupleIQs.R
\docType{data}
\name{coupleIQs}
\alias{coupleIQs}
\title{Couple IQs Example data}
\format{A data frame with 10 rows and 3 variables:
\describe{
  \item{couple}{couple}
  \item{boy}{boy}
  \item{girl}{girl}
}}
\usage{
coupleIQs
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
