% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_food.R
\docType{data}
\name{data_food}
\alias{data_food}
\title{Food Example data}
\format{A data frame with 6 rows and 3 variables:
\describe{
  \item{green}{green}
  \item{red}{red}
  \item{blue}{blue}
}}
\usage{
data_food
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
