% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_memory2.R
\docType{data}
\name{data_memory}
\alias{data_memory}
\title{Memory Example data 2}
\format{A data frame with 5 rows and 4 variables:
\describe{
  \item{incidental_agree}{incidental_agree}
  \item{incidental_disagree}{incidental_disagree}
  \item{intentional_agree}{intentional_agree}
  \item{intentional_disagree}{intentional_disagree}
}}
\usage{
data_memory
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
