% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_undergrad.R
\docType{data}
\name{data_undergrad}
\alias{data_undergrad}
\title{Undergrad Example data}
\format{A data frame with 16 rows and 4 variables:
\describe{
  \item{class}{class}
  \item{humanities}{humanities}
  \item{science}{science}
  \item{business}{business}
}}
\usage{
data_undergrad
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
