% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wait.R
\docType{data}
\name{data_wait}
\alias{data_wait}
\title{Wait Example data}
\format{A data frame with 7 rows and 3 variables:
\describe{
  \item{child}{child}
  \item{woman}{woman}
  \item{man}{man}
}}
\usage{
data_wait
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
