% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_GRE.R
\docType{data}
\name{GRE}
\alias{GRE}
\title{GRE Example data}
\format{A data frame with 5 rows and 3 variables:
\describe{
  \item{id}{id}
  \item{verbalGRE_1}{verbalGRE_1}
  \item{verbalGRE_2}{verbalGRE_2}
}}
\usage{
GRE
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
