% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_memory.R
\docType{data}
\name{memory}
\alias{memory}
\title{Memory Example data}
\format{A data frame with 9 rows and 3 variables:
\describe{
  \item{id}{id}
  \item{sound}{sound}
  \item{look}{look}
}}
\usage{
memory
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
