% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_memory_wide.R
\docType{data}
\name{memory_wide}
\alias{memory_wide}
\title{Memory Wide Example data}
\format{A data frame with 6 rows and 4 variables:
\describe{
  \item{id}{id}
  \item{digit}{digit}
  \item{letter}{letter}
  \item{mixed}{mixed}
}}
\usage{
memory_wide
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
