% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_react_wealth.R
\docType{data}
\name{react_wealth}
\alias{react_wealth}
\title{React Wealth Example data}
\format{A 3 x 3 table with rows of ignores, talks, and helps
\describe{
  \item{poor}{poor}
  \item{middle}{middle}
  \item{wealthy}{wealthy}
}}
\usage{
react_wealth
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
