% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_schizo.R
\docType{data}
\name{schizo}
\alias{schizo}
\title{Schizophrenia Example data}
\format{A data frame with 10 rows and 3 variables:
\describe{
  \item{id}{id}
  \item{yr_hos}{yr_hoss}
  \item{ori_test}{ori_test}
}}
\usage{
schizo
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
