% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_schizo_metal.R
\docType{data}
\name{schizo_metal}
\alias{schizo_metal}
\title{Schizophrenia Metal Example data}
\format{A data frame with 9 rows and 3 variables:
\describe{
  \item{patient}{patient}
  \item{year_2}{year_2}
  \item{year_3}{year_3}
}}
\usage{
schizo_metal
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
