% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_speed_voice.R
\docType{data}
\name{speed_voice}
\alias{speed_voice}
\title{Speed Voice Example data}
\format{A 2 x 3 table with rows of fast and slow
\describe{
  \item{child}{child}
  \item{woman}{woman}
  \item{man}{man}
}}
\usage{
speed_voice
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
