% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tasks_wide.R
\docType{data}
\name{tasks_wide}
\alias{tasks_wide}
\title{Tasks Wide Example data}
\format{A data frame with 5 rows and 6 variables:
\describe{
  \item{clerical_background}{clerical_background}
  \item{clerical_popular}{clerical_popular}
  \item{clerical_metal}{clerical_metal}
  \item{mechanical_background}{mechanical_background}
  \item{mechanical_popular}{mechanical_popular}
  \item{mechanical_metal}{mechanical_metal}
}}
\usage{
tasks_wide
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
