% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_test_scores.R
\docType{data}
\name{test_scores}
\alias{test_scores}
\title{Test Scores Example data}
\format{A data frame with 12 rows and 3 variables:
\describe{
  \item{id}{id}
  \item{spatial}{spatial}
  \item{math}{math}
}}
\usage{
test_scores
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
