% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_textbook_wide.R
\docType{data}
\name{textbook_wide}
\alias{textbook_wide}
\title{Textbook Wide Example data}
\format{A data frame with 9 rows and 5 variables:
\describe{
  \item{block}{block}
  \item{A}{A}
  \item{B}{B}
  \item{C}{C}
  \item{D}{D}
}}
\usage{
textbook_wide
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
