% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_woman_parents.R
\docType{data}
\name{woman_parents}
\alias{woman_parents}
\title{Woman Parents Example data}
\format{A 2 x 2 table with rows of self_divorced and self_married.
\describe{
  \item{home_broken}{home_broken}
  \item{home_complete}{home_complete}
}}
\usage{
woman_parents
}
\description{
Example data set from Cohen's book.
}
\keyword{datasets}
